Fix polygon winding order
I noticed an odd rendering issue with the vector tiles generated using the MapboxTileWriter. Random polygons where not rendering in Mapbox and when I extruded any polygons, they were missing faces. After a full day of debugging my app I track the issue down to a small bug in this project. The issue is with the winding order of the polygons. The [Mapbox Vector Tile Specification](https://docs.mapbox.com/vector-tiles/specification/) says that the shell ring should be CW and holes be CWW oriented. This is the opposite of the OGC standard which is likely where the confusion came from.