Add support for Feature IDs (#3)
NetTopologySuite features don't have an ID property. As such, this library isn't currently assigning any ID value to any features when writing tiles. This creates some limitations when using the tiles elsewhere (i.e. more difficult to use feature states).

This PR adds a support for specifying which attribute should be used as an ID for the feature in the tile. This is fully backwards compatible. By default it will look for an attribute called "id" (you can pass in a different value to use). If no attribute with the specified key is found, it simply won't assign any ID. If it does find a value, it then tries to parse it as a ulong, and if successful, passes the value in as the ID on the feature.

Added tests and reader support