Update TileGeometryTransform.cs
Reverting change. This was not a part of my PR. It came in when I went to upload my original changes and merged down from somewhere. I'm assuming the original author of this bit of code added the comments that the errors in question should never happen and thus the code is not used.